Add support for an optional auto-historical check box
When the check box is checked, new residential and commercial lots will
be marked as historical when the game creates them.
This is intended as a convenience for people who are using the building
styles to make SC4 construct neighborhoods with a specific theme.
The historical status of existing lots will not be affected by the
auto-historical check box, there are existing mods that can be used to
get that behavior (e.g. the AutoHistorical DLL).

The state of the checkbox will be persisted to a new DBPF record in the
city save file, this record will be ignored by SC4 itself.