Add a cIBuildingStyleInfo GZCOM interface
This allows other DLLs to query the available building style ids and
get the style names.

It also provides a function to get the style names that are supported
by a specific building occupant as a comma-separated string, this
replaces the custom code that the query UI extensions DLL used for that
purpose.
The query UI extensions DLL code was broken by the addition of the
auto-historical check box, which it would mistake for a style id.
Moving that functionality into the more building styles DLL should
prevent that problem from repeating in the future.