[WIP] Hook SC4's building style selection code
This is a starting point for replacing that code with a custom
implementation, as well as further investigations into its behavior.

Note that trying to use the cISC4LotConfiguration pointer crashes.
This is an issue of retrieving the correct pointer from the CPU
registers, but I am not sure what the problem is.

Looking at the assembly, EDI should be the register that contains the
correct pointer, but it crashes when trying to use it.