Fix the C++ lot style selection hook code
The MSVC compiler was triggering a write access violation with the
assembly code it was producing as part of a call to
cISC4LotConfiguration::IsCompatibleWithStyleType.
The access violation occurred when trying to write data onto the
stack.

As a workaround we now use a LotConfigurationHasStyle method that
recreates what IsCompatibleWithStyleType would be doing by directly
accessing the cSC4LotConfiguration object's occupant group vector.