Read the building type from the EBP register
SC4 is using EBP as a general purpose register instead of for storing
the stack base address.

But the hook crashes due to EBP not being restored to an address the
game expects to be able to jump to when the hook returns.
The game code calls pop EBP before the remaining code is run, but that
doesn't work in this case.