Expand the BuildingStyles.ini format
It now allows disabled placeholder check boxes and choosing between
normal and bold text.
Any unused check box that is not a placeholder will be hidden.

A style id value of 0 is now reserved as a marker for the placeholder
styles.
Some Maxis blockers are implemented using style id 0, so we disallow
it from being used as a building style id.

The LTEXT format string was changed to CaptionRes, and the type id
value was dropped due to it being redundant.