Add a GZCOM class that provides the wall-to-wall info
The DLL identifies wall-to-wall (W2W) buildings by the presence of
specific occupant groups.
The cIBuildingStyleWallToWall interface allows other GZCOM DLLs to
get a list of the supported W2W occupant groups, get the group names,
and check if a specific building is part of one or more of those
occupant groups.