Add support for a Kick Out Lower Wealth check box
This allows the user to toggle whether the tract developer will kick
out lower wealth occupants when redeveloping.

The default value will be based on the user's installed plugins.
If the No Kick Out mod or a similar mod is installed the default value
will be false, otherwise it will be true.