Add a BuildingStyles exemplar property
If this property is present, the building will only grow under those
styles and not any of the 4 Maxis styles.
The BuildingStyles property is also a prerequisite for a future rewrite
of the industrial building style support.

Added a LotConfigurationManager hook to make SC4 include the data in
the BuildingStyles property when it fetches the lot occupant groups.