Allow the Building Styles property to use Maxis styles
When the Building Styles property is present, the only items that will
be copied from the occupant groups property is the wall-to-wall style
data, if present.

Because the original building styles are no longer copied over from the
occupant groups property, the block on the Maxis styles is no longer
required.