Add logging for the game's candidate lots
Candidate lots are existing lots or empty RCI zones that the game picks
to (re)develop.
We log various details, lot cell positions, size, developer, etc.
The existing building name will be logged if present, it will not
be available for an empty lot or a building that is still under
construction.