Add support for a Prevent Cross-Style Redevelopment check box
When checked, it attempts to force the game to replace existing
buildings that it redevelops with one that has the same building style.

It works, but I am not sure if it is 100% reliable. There are cases
where SC4 will skip the existing building style check, e.g. if the
building/lot doesn't exist.
But I don't know how common those cases are.