Expand Adding More Building Styles into SC4
This section has been reorganized to clearly separate the 2 modding
areas the DLL's documantation covers.

1. Modifying SC4's Building Style UI to support custom building styles.
2. Adding custom style support to the exemplars of existing buildings.

A copy of the CoriBoom Building Style UI template that uses
BuildingStyles.ini is now included in the repository, and the
documentation for the new check boxes and radio buttons that the DLL
supports has been moved to its own file to simplify the main readme.