Special case the PIM-X placeholder style id
PIM-X uses the value 0x7b6bc069 as a placeholder in its Building Styles
property template, where the mod author is meant to replace it with a
real style id.

Unfortunately a large number of buildings were updated to use this
placeholder value, which would act as a blocker when the DLL is
installed.
Additionally, there are a number of other values that can't be used as
a style id, e.g. the control ids in the Building Style Control UI.

To fix these issues, a number of values have been reserved as off
limits for style ids, this does not affect any values in the current
0x2000-0x2FFF range used for building styles.

Buildings that have only the reserved style ids in their
Building Styles property will be made to use the legacy Maxis
styles in the Occupant Groups property.

Closes #1