Add a way to register styles using an exemplar
This is better for mods that are distributed on the exchanges than
requiring users to manually edit BuildingStyles.ini to add the new
styles to the user interface.

The building style exemplar must use the group id 0xB06361D6 with the
style id as the instance id.
The Building Style Exemplar: Style Name Key property is mandatory, but
other properties can also be set to configure the use of bold text
and/or add a tool tip for the style.

The styles specified in BuildingStyles.ini take precedence over any
conflicting styles specified in an exemplar.
This is done to preserve the behavior of the INI file allowing for
fully custom styles.