Fix buildings that don't have any UI styles
If a building exemplar has all of its Building Styles property values
set to styles that are not present in the Building Style Control UI,
the Building Styles property will be ignored and the building will use
the Maxis styles from the occupant groups property.

An error message will also be written to the log file identifying the
problematic building exemplar.

The above scenario should be uncommon for users who are downloading
buildings with custom building styles, as their dependencies should
include the updated Building Style Control UI and an exemplar to add
the style to the UI.

As a side effect, this will allow the Building Styles property to be
bypassed by users who are only using the DLL for its fix to
Build all styles at once.
But at the cost of a bunch of log messages and some performance loss
from the additional checks, so those users would be better served by
removing the Building Styles property from their downloaded mods.