Fix handling of the PIMX 0x2004 placeholder style
Some versions of PIMX accidentally used the community style id 0x2004
as a placeholder in the Building Styles property.

For currently released buildings that use that placeholder, we fall
back to the Maxis styles.
Buildings released in the future can use 0x2004 as an actual style, a
marker property was added to the PIMX templates which allows that case
to be distinguished from the placeholder usage.

PIMX building exemplars are detected by the presence of the building
template id in the ExemplarCategory property.